refactor: interceptor, configuration 패키지 분리
