+++ src/main/java/techcourse/myblog/configuration/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web;
+package techcourse.myblog.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
