+++ src/main/java/techcourse/myblog/configuration/WebMvcConfig.java
@@ -1,9 +1,11 @@
-package techcourse.myblog.web;
+package techcourse.myblog.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import techcourse.myblog.interceptor.LoginInterceptor;
+import techcourse.myblog.interceptor.LogoutInterceptor;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
